requirements.txt
langchain
pypdf
sentence-transformers
faiss-cpu
ctransformers